requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
psycopg2-binary
sqlalchemy
cryptography
fastapi
uvicorn
pydantic
streamlit-lottie
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
pytest-cov
httpx
supabase>=2.0.0
postgrest-py
